// mini-xrm-frontend/mini-xrm-client-angular/model/personResponsibleReportSortFieldView.ts
/**
 * Mini xRM API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const PersonResponsibleReportSortFieldView = {
    PersonResponsible: 'personResponsible',
    TotalDurationMinutes: 'totalDurationMinutes',
    PartnerCount: 'partnerCount'
} as const;
export type PersonResponsibleReportSortFieldView = typeof PersonResponsibleReportSortFieldView[keyof typeof PersonResponsibleReportSortFieldView];