// mini-xrm-frontend/mini-xrm-client-angular/model/sortDirectionView.ts
/**
 * Mini xRM API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const SortDirectionView = {
    Asc: 'asc',
    Desc: 'desc'
} as const;
export type SortDirectionView = typeof SortDirectionView[keyof typeof SortDirectionView];